+++ packages/biscuit-store/src/internal/helper.js
@@ -2,18 +2,34 @@ import { middlewares, states, repositories } from './repositories';
 import { CreateError } from './debugger';
 import { messages } from './messages';
 
-export function getStateRepo(action) {
+/**
+* Get store link
+* @param {string} name
+*/
+export function getStateLink(action) {
     return states[`"${action.type}"`][action.name];
 }
 
-export function getStoresitory(name) {
+/**
+* Get store content
+* @param {string} name
+*/
+export function getStoreContent(name) {
     return repositories[name].content;
 }
 
-export function getStoresitoryActions(name) {
+/**
+* Get store actions
+* @param {string} name
+*/
+export function getStoreContentActions(name) {
     return repositories[name].actions;
 }
 
+/**
+* To obtain the name of the store depending on the type of
+* @param {object | string} target
+*/
 export function getStoreName(target) {
     if (typeof target === 'string') {
         return target;
@@ -22,6 +38,10 @@ export function getStoreName(target) {
     return target.name;
 }
 
+/**
+* Validating an action
+* @param {import('../../types/state').AnyAction} action
+*/
 export const actionError = (action) => {
     if (!action || !action.name || !action.type) {
         throw new CreateError('Invalid action parameters.');
@@ -68,43 +88,38 @@ export function gettter(instance) {
 /**
  * Helper method for comparing two objects
  * Warning: can't compare methods
- * @param {object} firstState first object
- * @param {object} lastState last object
+ * @param {object} obj1 first object
+ * @param {object} obj2 last object
  * @return {bool}
  * @private
  */
-export function compareObject(firstState, lastState) {
-    let propInFirst = 0;
-    let propInLast = 0;
-    let prop;
-
-    if (firstState === lastState) {
+export function compareObject(a, b) {
+    if (a === b) {
         return true;
     }
 
-    if (
-        firstState === null ||
-		typeof firstState !== 'object' ||
-		lastState === null ||
-		typeof lastState !== 'object'
-    ) {
+    if (a === null || typeof a !== 'object' || b === null && typeof b !== 'object') {
         return false;
     }
 
-    for (prop in firstState) {
-        propInFirst += 1;
+    if (Object.keys(a).length !== Object.keys(b).length) {
+        return false;
     }
 
-    for (prop in lastState) {
-        propInLast += 1;
-
-        if (
-            !(prop in firstState) ||
-			!compareObject(firstState[prop], lastState[prop])
-        ) {
-            return false;
+    let equal = true;
+    for (let key in a) {
+        if (typeof a[key] === 'object' && typeof b[key] === 'object') {
+            if (!compareObject(a[key], b[key])) {
+                equal = false;
+            }
+        } else if (typeof a[key] === 'function' && typeof b[key] === 'function') {
+            if (a.toString() !== b.toString()) {
+                equal = false;
+            }
+        } else if (a[key] !== b[key]) {
+            equal = false;
         }
     }
 
-    return propInFirst === propInLast;
+    return equal;
 }