+++ CHANGELOG.md
@@ -7,4 +7,4 @@
 - Added call method to adapter
 - Added channel method to adapter
 - Added createActionTo unit test
-- refactor store adn actions
+- refactor store and actions